+++ stack--2022/2-foundation/error-utils/src/fields.ts
@@ -31,6 +31,9 @@ export const COMMON_ERROR_FIELDS = new Set<keyof XError>([
 	'code', // https://medium.com/the-node-js-collection/node-js-errors-changes-you-need-to-know-about-dc8c82417f65
 	        // https://nodejs.org/dist/latest/docs/api/errors.html#errors_node_js_error_codes
 
+	// standard in Firefox only for AggregateError:
+	'errors', // https://devdocs.io/javascript/global_objects/aggregateerror
+
 	// non standard but widely used:
 	'statusCode', // express, https://gist.github.com/zcaceres/2854ef613751563a3b506fabce4501fd
 	'shouldRedirect', // express, https://gist.github.com/zcaceres/2854ef613751563a3b506fabce4501fd